backend/requirements.txt
django~=4.2.1
Pillow
djangorestframework~=3.14.0
django-redis
django-q
django-cors-headers
environs